+++ nanoutils/__init__.py
@@ -4,12 +4,13 @@
 
 from .__version__ import __version__
 
-from . import typing_utils, empty, utils, numpy_utils, schema, set_attr
+from . import typing_utils, empty, utils, numpy_utils, schema, set_attr, file_container
 from .typing_utils import *
 from .empty import *
 from .utils import *
 from .numpy_utils import *
 from .set_attr import *
+from .file_container import *
 from .schema import (
     Default, Formatter, supports_float, supports_int,
     isinstance_factory, issubclass_factory, import_factory
@@ -26,3 +27,4 @@ __all__ += utils.__all__
 __all__ += numpy_utils.__all__
 __all__ += schema.__all__
 __all__ += set_attr.__all__
+__all__ += file_container.__all__